fix: Resolve TypeScript build errors for Vercel deployment
- Add type assertion for Uint8Array in useAudioSystem.ts (line 106)
- Prefix unused parameter with underscore in useVideoRecorder.ts (line 156)
- Comment out unused angleZ variable in MentalFocus.tsx (line 73)

+++ src/hooks/useAudioSystem.ts
@@ -103,7 +103,7 @@ export const useAudioSystem = () => {
       return null;
     }
 
-    analyserRef.current.getByteFrequencyData(dataArrayRef.current);
+    analyserRef.current.getByteFrequencyData(dataArrayRef.current as Uint8Array);
     const data = dataArrayRef.current;
 
     // Bass: 0-50, Mid: 50-150, High: 150-255
